Refactor core structure related with package
