+++ judge-core/src/lib.rs
@@ -1,9 +1,9 @@
-pub mod builder;
 /// Helper for compiling source into an executable
 pub mod compiler;
 pub mod error;
 /// Prepared functions for running one single judge
 pub mod judge;
+pub mod package;
 /// Runtime essentials for running a program under control
 pub mod run;
 pub mod utils;
